Only read next line if stdin is not a tty
When reading from a pipe, it probably makes sense to read the next line,
even if it blocks for a second. But when reading from a terminal, this
is probably just confusing for users. We’ve printed the error, the user
probably won’t type anything else: let’s not wait for any further input.